t/t-migrate-import.sh: fixup grep with check-attr
In commit 8490bf31483a06db435774358373053e3fcee5fd of PR #4512 we
updated the "git lfs migrate import" command so that when its --above
option is specified, the command only adds entries for individual
files to the ".gitattributes" rather than adding entries with wildcard
patterns.

To verify these changes, we also added one test and updated two others
in our t/t-migrate-import.sh test script.  In each of these tests, we
added a final check that the "filter" Git attribute is not set to the
value "lfs" for a file that is below the threshold size specified with
the "git lfs migrate import" command's --above option.

These checks execute the "git check-attr" command and supply both the
attribute name "filter" and a file path, and pipe the command's output
to the grep(1) command.  The "grep" command is invoked with its -v option,
so it filters from the output of the "git check-attr" command any lines
that match its pattern argument, which is the string "lfs".

In several prior commits in this PR we addressed the instances in our
shell test suite where we used the "grep" command with the invalid
assumption that the command would fail if one of its input lines matched
the pattern provided with the -v option.

While the checks added in commit 8490bf31483a06db435774358373053e3fcee5fd
also use the "grep" command's -v option, they function as intended,
but for a somewhat subtle reason.

When the "git check-attr" command is run and passed a single attribute
name and a single file path, it will always output a single line, even
if the attribute is not defined for the given file.  The Git project's
documentation specifies the format of this line, which includes the
file path, the attribute name, and an "info" field, separated by colon
and space characters:

  https://git-scm.com/docs/git-check-attr#_output

If the attribute has a specific value for the given file, the "info"
field contains the value, and if the attribute is not defined for the
file, the "info" field contains the string "unspecified".

As a consequence, when the tests in the t/t-migrate-import.sh script
pass the output of the "git check-attr" command to the "grep" command,
the latter is guaranteed to always have one line of input.  Further,
because we expect the "filter" attribute to not be defined for the
given files, this line will not contain the string "lfs" but instead
the string "unspecified", and so the "grep" command will not filter
its input line because it doesn't match the pattern provided with
the -v option.  The "grep" command therefore returns an exit status
code of zero, and so our tests succeed.

While these "grep" commands with the -v option are successful and
function as intended, the reason for their behaviour remains somewhat
obscure.  We therefore revise them now to remove the -v option and
replace their pattern arguments with the string "filter: unspecified",
which makes more explicit the output they expect from the
"git check-attr" command.

We also revise two other instances where we pipe the output of the
"git check-attr" command to the "grep" command, both of which were
added to the "migrate import (filename special characters)" test in
the t/t-migrate-import.sh script in commit
5d7f6ec3f36a3191bbbc741d4c6a2240b430c95f of PR #4683.

In these two cases, we expect the "git check-attr" command to report
that the "filter" attribute has the value "lfs" for each file matching
a shell wildcard pattern.  At present, we pipe the output of the command
to a "grep" command with the pattern argument "lfs", and then pipe that
command's output to the wc(1) command to count the number of lines
containing the string "lfs".  This line count value is then compared
to the number of files which should match the wildcard pattern to ensure
they are all assigned the expected "filter" attribute value of "lfs".

These two checks are both successful and function as intended.  However,
in previous commits in this PR we have revised many other tests which
instead use the "grep" command's -c option to count the number of lines
from its input which match a given pattern.

To increase the consistency between our tests, we therefore now also
revise the two checks in the "migrate import (filename special characters)"
test by using the "grep" command's -c option to count the lines of input
matching the given pattern.  We also expand the command's pattern argument
to the string "filter: lfs", which makes more explicit the output we
expect from the "git check-attr" command in these tests, and we reorder
the enclosing "[" shell test builtin expressions so that the expected
line count value always appears as the first operand.

All of these changes both simplify the tests in the t/t-migrate-import.sh
script and help bring them into alignment with the many similar commands
throughout our test suite that we have revised or corrected in prior
commits in this PR.