t/testhelpers.sh: simplify attr count with grep -c
Since commit 3652681458953cee47ee3c028e15a9e4893c52f4 of PR #1870 the
assert_attributes_count() assertion function in our shell test library
has used the grep(1) and wc(1) commands to count the number of lines in
the ".gitattributes" file which match a given pattern.

We now simplify this operation by using the -c option of the "grep"
command to count the number of matching lines instead of a separate
"wc" command.  We also replace the "!=" comparison operator in the
subsequent shell test with the -eq operator, since we expect both of
its operands to be integers.

These changes slightly simplify the steps in the assert_attributes_count()
function, and also bring them into alignment with the many similar commands
throughout our test suite that we have revised or corrected in prior
commits in this PR.

Note that unlike those other instances where we make use of the "grep"
command's -c option, in this case the command runs in a subshell within
a simple assignment rather than within a "[" shell builtin, so we need
to append a "true" command following an "||" OR list operator.

This is required to handle the case where the "grep" command finds no
lines in its input which match the given pattern.  The command will
output a line count of "0" to its standard error file descriptor, but
will also return an exit code of 1, indicating that the command did not
find any matches.  Our tests always enable the "errexit" shell attribute
using the "set -e" command, and so this would cause the test which called
the assert_attributes_count() to fail, because the "grep" command's exit
status would be treated as the status of the simple variable assignment.
We therefore need to append an "||" operator and a "true" command so
that the subshell will always return a zero exit status.

By contrast, all our other uses of the "grep" command with the -c option
occur in subshells within "[" shell builtins, and so the commands' exit
statuses do not supplant the statuses of the "[" builtins themselves.