t/t-credentials.sh: simplify checks with grep -c
A number of tests in our t/t-credentials.sh test script perform Git LFS
operations and capture their trace log messages into a log file, and
then check that a certain number of lines in the log file match a
given pattern.

At present, many of these checks use a technique first introduced in
commit d340a739d97abdadc3182c8612ff1221f53c689c of PR #2695 whereby the
log file's contents are read with the cat(1) command and piped to the
grep(1) command, and then the matching lines output by the "grep" command
are piped to the wc(1) command to be counted.

We now simplify these checks by replacing the three commands with a
single "grep" command, using the -c option to count the lines of input
matching the given pattern, and passing the log file's name as an
argument to the "grep" command instead of using the "cat" command to
read the file.  We remove the quotation marks around the integers
used in the surrounding "[" shell builtins, since our checks use
the -eq shell test operator and it expects integer operands.

These changes both simplify the tests in the t/t-credentials.sh script
and help bring them into alignment with the many similar commands
throughout our test suite that we have revised or corrected in prior
commits in this PR.