docs/man/git-lfs-merge-driver.adoc: add output arg
In commit 402e958bfa121818dabb0aba2ff7b6865dfe9dbf of PR #4970 we
implemented the new "git lfs merge-driver" command, which can be
configured as a custom Git merge driver in order to help automatically
merge some types of files when they are tracked as Git LFS objects.

We expect this command to be invoked by Git during a merge operation,
and so several of its command-line flags are required, as their arguments
must specify the paths of the temporary files created by Git for use
by a merge driver program.

Our git-lfs-merge-driver(1) manual page describes these flags but
accidentally omits one of them, the --output flag, so we now add an
entry for it into the list of options supported by the command.

We also take the opportunity to slightly adjust the description of the
"%D" placeholder token to indicate that it corresponds to the value of
the --output option's argument.  (If the user specifies a custom command
with the --program option, the "git lfs merge-driver" command will run
that command after substituting any placeholder tokens with their
corresponding values.)

Note that technically some of the "git lfs merge-driver" command's flags
and their arguments are not actually options at all, since they are
required for the "git lfs merge-driver" command to run.  This includes
the --output flag as well as the --ancestor, --current, and --other
flags.

However, since we expect to further revise the git-lfs-merge-driver(1)
manual page in a future PR, we confine ourselves at present to simply
adding an entry for the missing --output flag to the existing list of
options and clarifying the flag's corresponding "%D" placeholder token.