Makefile,script: add vendored source release asset
As described in #6071, some of the downstream distributors who package
Git LFS from our source release asset also run their build processes
offline and therefore they cannot run "go mod vendor" to fetch our
Go module dependencies.  As a result, these distributors have requested
that we also provide a source release asset which includes a
fully-populated "vendor" directory.

To fulfill this request, we add a new release asset target to our
Makefile and include this new asset's archive file in the list in the
"RELEASE_TARGETS" make(1) variable.  The new target is named
"bin/releases/git-lfs-vendor-$(VERSION).tar.gz", and contains both
our source files and a complete "vendor" directory.

We build this new release asset in several steps.  First, we run
"git archive" akin to how it is used in our regular source release
asset target, the "bin/releases/git-lfs-$(VERSION).tar.gz" target,
except that we generate a local "tmp.tar" file rather than a final,
compressed archive file.

We then add the contents of the "vendor" directory to this temporary
archive file, using the appropriate command-line arguments to rewrite
the names of the added files so they include our expected versioned
package name as a parent directory.  This ensures the "vendor" directory
is located alongside our source files in the same directory when the
archive file is unpacked.

Note that the "vendor" directory must exist in the current working
directory at the time our new Makefile target's recipe runs because we
specify the "vendor" prerequisite for the target.  Our existing "vendor"
target runs "go mod vendor", so this will create and populate the
"vendor" directory before our new target's recipe executes.

We do, however, ensure that the files in the "vendor" directory have
their group permissions set, to match the permissions of the files in
the temporary archive files.  The "git archive" command uses a default
octal permissions mask of 0002 when adding files to the archive it
outputs, so we adjust our "vendor" directories to align with that default:

  https://git-scm.com/docs/git-archive#Documentation/git-archive.txt-tarumask

Finally, we compress the archive using the gzip(1) command to create
our final release asset file, and remove the temporary archive file.
To invoke the "gzip" command we reference a new "GZIP" variable in our
Makefile whose default value is "gzip"; this aligns with our use of similar
variables for the grep(1), tar(1), and xargs(1) commands.  (We do not
consistently reference these "make" variables, however, but we defer
addressing that issue to a future PR).

In addition to revising our Makefile, we also update our "script/upload"
file so that its categorize_asset() function recognizes our new release
asset archive file and provides a description for it, to be used in the
asset file listings of our GitHub release pages.