docs/man: remove underscores in synopsis arguments
The documentation of our commands generally conforms to the conventions
of the Unix and Git manuals.  However, a number of inconsistencies are
present in both the content and format of our manual pages.

In a previous commit in this PR we revised all the synopsis sections
of our AsciiDoc manual pages so they were formatted consistently using
source listing blocks, which helps us avoid problems with missing
hard line breaks, among other advantages.

In a subsequent commit in this PR we expect to make additional changes so
that when our manual pages are rendered into HTML or roff, either by an
Asciidoctor converter or by GitHub's UI, the synopsis sections of our
pages will identify literals using boldface text and replaceable arguments
using italics.  This will bring our documentation into closer alignment
with the standards of the Linux man-pages project:

  https://git.kernel.org/pub/scm/docs/man-pages/man-pages.git/tree/man/man7/man-pages.7?id=a5342ef55f0a96790bf279a98c9d2a30b19fc9eb#n200

However, our "mangen" program should continue to render our manual pages
into plain text, as this is the format in which they are output by the
"git lfs help" command.  We will therefore update this program to remove
the AsciiDoc markup for boldface and italic text from the content of the
synopsis sections of our manual pages.

Before we make these changes, we first need to revise the synopsis
sections of several manual pages whose existing format would conflict
with the additional AsciiDoc markup we intend to use to identify
italic text in particular.

AsciiDoc uses underscore and asterisk characters to demarcate italic and
boldface text, respectively.  Since we expect to make use of these
formatting characters, we will also have to update our "mangen" program
so that it removes them from our synopsis sections.

Ultimately, we intend to only use the underscore and asterisk characters
in constrained pairs rather than unconstrained pairs, meaning they should
only appear immediately before or after terms, never within a term:

  https://docs.asciidoctor.org/asciidoc/latest/text/#constrained
  https://docs.asciidoctor.org/asciidoc/latest/text/#unconstrained

However, adjusting our "mangen" program to remove underscore and
asterisk characters from synopsis sections will be substantially
simpler if we do not have to account for this difference.  We can just
remove all such characters from our synopsis sections instead of trying
to match formatting characters only when they appear between the start
or end of a term and either whitespace or the start or end of a line.

For this reason, we would prefer that no underscore characters appear
in any terms in our synopsis sections, as they do now in our
git-lfs-post-checkout(1) and git-lfs-post-merge(1) manual pages.
At present, we define the replaceable arguments of those two commands with
terms that contain underscore characters, such as "is_branch_checkout"
and "is_squash".

As a further incentive to replace the underscore characters in these
two manual pages' synopsis sections, the Git project's guidelines for
manual page documentation state that command examples should use only
"dashes" (i.e., the ASCII "hyphen-minus" character) to separate words:

  https://github.com/git/git/blob/c4a0c8845e2426375ad257b6c221a3a7d92ecfda/Documentation/CodingGuidelines#L871-L878

Therefore, both to bring our documentation into closer conformance with
Git's standards and to allow our changes to the "mangen" program to be
as simple as possible, we change the underscore characters in the
git-lfs-post-checkout(1) and git-lfs-post-merge(1) manual pages'
synopsis sections to hyphen characters.