docs/man,t: add formatting to synopsis sections
In previous commits in this PR we revised all the synopsis sections of
our AsciiDoc manual pages so they are formatted consistently using source
listing blocks, and ensured that no AsciiDoc inline formatting characters
appear in the synopsis sections.

With these changes in place we are now able to reformat our synopsis
sections so that when our manual pages are rendered into HTML or roff,
either by an Asciidoctor converter or by GitHub's UI, our synopses will
consistently identify literals using boldface text and replaceable
arguments using italics.  In doing so, we will bring our documentation
into closer alignment with the standards of the Linux man-pages project:

  https://git.kernel.org/pub/scm/docs/man-pages/man-pages.git/tree/man/man7/man-pages.7?id=a5342ef55f0a96790bf279a98c9d2a30b19fc9eb#n200

First, we apply the AsciiDoc "quotes" substitution rule to all of our
synopsis sections, so that Asciidoctor will respect the AsciiDoc inline
formatting characters we want to use, namely the underscore and asterisk
characters for boldface and italic text, respectively.  We also need to
specify the "verbatim" rule, since we are overriding the default rule
settings for source listing blocks, but still want to start by applying
that rule:

  https://docs.asciidoctor.org/asciidoc/latest/subs/#verbatim-group
  https://docs.asciidoctor.org/asciidoc/latest/subs/quotes/

We then add the appropriate inline formatting markup to all of the
synopsis sections of our manual page source files.  We demarcate all
literals using boldface text, and all replaceable arguments using italics.

Note that in the our git-lfs-pointer(1) manual page we make use of
unconstrained formatting pairs in order to render the existing synopsis
text without other changes, as this page is the only one where we include
the optional equals sign character between literal and replaceable
arguments, e.g., "--file=path/to/file" instead of "--file path/to/file".
We expect to further revise this page in a subsequent PR at which time we
will also rewrite a number of other pages' synopsis sections so they use
a consistent style for all of their text and placeholder arguments, and
only require the use of constrained formatting pairs:

  https://docs.asciidoctor.org/asciidoc/latest/text/#constrained
  https://docs.asciidoctor.org/asciidoc/latest/text/#unconstrained

Next, we revise our "mangen" program so that it continues to render
our manual pages into plain text, as this is the format in which they are
output by the "git lfs help" command.  Specifically, we update this
program to remove the AsciiDoc markup for boldface and italic text from
the content of the synopsis sections of our manual pages.

The "mangen" program has recognized the start and end of AsciiDoc source
listing blocks since commit f11d3099ae27a6bea5ebc8beada9ac6978a50a0d of
PR #5054, when we first migrated our manual pages' source files to the
AsciiDoc format.  We now expand this logic so that it detects when a
source listing block is defined with our custom "synopsis" role and
sets an internal "isSynopsisSourceBlock" flag, which remains "true" until
the block's final line is encountered.

While the "isSynopsisSourceBlock" flag is set we apply a simple regular
expression which removes any asterisk or underscore characters from
each line.  As we noted in a previous commit in this PR, this approach
is significantly less complicated than if we attempted to match inline
formatting characters only when they appear between the start or end of
a term and either whitespace or the start or end of a line.  Further, this
regular expression also allows us to handle unconstrained formatting
pairs, which as explained above we make use of in our git-lfs-pointer(1)
manual page for now.

Finally, we introduce a new t/t-usage.sh shell test script with a single
initial "usage: no command specified" test which verifies that our
git-lfs(1) manual page is output by the multiple variants of the "git lfs"
and "git-lfs" commands.  Note that we cannot test using the "git lfs -h"
or "git-lfs -h" variants because at present we do not support the -h option
when it is provided without a Git LFS command name, as discussed in #6168.